requirements.txt
click==6.7
codecov==2.0.15
coverage==4.5.1
pychalk==2.0.0
pyyaml==3.12
pathlib2==2.3.0